removed abstraction and reinstated rds Storage
